convert solution to vs2017 compatible format and fix nuget package targets
The package generated by the nuspec file previously included targeted net45 and netcore, though the TwoStepsAuthenticator.DotNetCore targeted netstandard1.6. this led to (lib) projects targeting netstandard could not consume the package, even though the lib would support it correctly.

After converting to VS2017 (autoconverted, by VS itself) the following was done:
* Changed TwoStepsAuthenticator.DotNetCore.csproj to target both net45 and netstandard1.6. The new csproj format allows for multitargeting and package generation from it, so the two projects and the .nuspec are no longer required
** added package generation info into the csproj
** removed the TwoStepsAuthenticator project
** renamded TwoStepsAuthenticator.DotnetCore to TwoStepsAuthenticator
** removed the nuspec file
* package can be generated from the root with a `dotnet pack -c Release` command and will be stored in `TwoStepsAuthenticator\bin\Release`. Other configs are possible from the build server.

The unit test project targets net45 and netcore1.0 to run the tests for both configurations. Nunit was upgraded to 3.9. Tests can be run using `dotnet test`from the unit test project folder

Notes:
* UsedCodesManager had some differences between the net45 and netstandard projects (it was the only non linked file) though it seems to be ok from the test run.
* If signing is useful, that can also be added to the csproj or just the CI build
* package set at 1.3.0
