+++ .gitignore
@@ -153,3 +153,6 @@ $RECYCLE.BIN/
 # Mac desktop service store files
 .DS_Store
 packages/
+
+#VS definitions folder
+.vs
