--- TwoStepsAuthenticator.DotnetCore/project.json
@@ -1,14 +1,0 @@
-{
-  "version": "1.0.0-*",
-  "dependencies": {
-    "NETStandard.Library": "1.6.0"
-  },
-  "frameworks": {
-    "netstandard1.6": {
-      "imports": "dnxcore50"
-    }
-  },
-  "buildOptions": {
-    "define": [ "CORE" ]
-  }
-}
