+++ TwoStepsAuthenticator.TestApp/TwoStepsAuthenticator.TestApp.csproj
@@ -96,7 +96,7 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\TwoStepsAuthenticator\TwoStepsAuthenticator.csproj">
-      <Project>{6c898cd1-0bf3-4711-847e-ad7dac815cd8}</Project>
+      <Project>{03842fee-34c4-467f-9081-96bed41cefb2}</Project>
       <Name>TwoStepsAuthenticator</Name>
     </ProjectReference>
   </ItemGroup>
