+++ TwoStepsAuthenticator.TestApp/ViewModel.cs
@@ -1,10 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Timers;
 using System.Windows.Threading;
 
 namespace TwoStepsAuthenticatorTestApp
